backend/requirements.txt
asgiref==3.7.2
pytz==2023.3.post1
sqlparse==0.4.4
django-extra-fields==3.0.2
django-filter==23.2
djoser==2.1.0
Pillow==9.3.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gunicorn==20.1.0
Django==3.2.16
djangorestframework==3.12.4
django-cors-headers==3.13.0